Extract AbstractVersionsStep.
